net: Move eviction node selection logic to SelectNodeToEvict(...)
